added typescript to about.tsx
